Branch Order

+++ hhn-calamba-be/hhn/src/main/java/com/ally/hhn/service/PromoService.java
@@ -1,5 +1,6 @@
 package com.ally.hhn.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.json.simple.JSONObject;
@@ -9,8 +10,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import com.ally.hhn.model.Item;
 import com.ally.hhn.model.Promo;
 import com.ally.hhn.model.PromoDTO;
+import com.ally.hhn.model.PromoJSON;
+import com.ally.hhn.model.PromoItem;
 import com.ally.hhn.repository.PromoItemRepository;
 import com.ally.hhn.repository.PromoRepository;
 
@@ -44,7 +48,7 @@ public class PromoService {
 		return json;
 	}
 	
-	public void save(PromoDTO promoDTO) {
+	public void save(PromoJSON promoDTO) {
 		final Promo savedPromo = promoRepository.save(promoDTO.getPromo());
 		
 		promoDTO.getPromoItemList().forEach(promoItem -> {
@@ -54,14 +58,25 @@ public class PromoService {
 	}
 	
 	public void delete(Promo promo) {
-		promoItemRepository.findByPromoId(promo.getPromoId()).forEach(promoItem -> {
-			promoItemRepository.delete(promoItem);
-		});
+		for(PromoItem promoItem: promoItemRepository.findByPromoId(promo.getPromoId())){
+			if(promoItem != null) promoItemRepository.delete(promoItem);
+		}
+
 		promoRepository.delete(promo);
 	}
 	
 	public void save(Promo promo) {
 		promo.setActive(!promo.isActive());
 		promoRepository.save(promo);
+	}
+	
+	public List<PromoDTO> getAllPromos(){
+		List<PromoDTO> promoDtoList = new ArrayList<PromoDTO>();
+		for(Promo promo: promoRepository.findAll()) {
+			List<PromoItem> promoItems = promoItemRepository.findByPromoId(promo.getPromoId());
+			PromoDTO promoDto = new PromoDTO(promo, promoItems);
+			promoDtoList.add(promoDto);		
+		}
+		return promoDtoList;
 	}
 }
